Siniflarin ayri dosyalara ayrilmasi.
